backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.16
faiss-cpu==1.7.4
pandas==1.3.3
requests==2.26.0
python-dotenv==0.19.0